+++ rest/src/index.js
@@ -69,7 +69,6 @@ const loadConfig = () => {
 	configFiles.forEach(configFile => {
 		winston.info(`loading config from ${configFile}`);
 		const partialConfig = JSON.parse(fs.readFileSync(configFile, 'utf8'));
-		console.log('AAAAAAAAAAAAAAAAAAAAAAAa',partialConfig)
 		if (config) {
 			// override config
 			catapult.utils.objects.checkSchemaAgainstTemplate(config, partialConfig);